==> src/nn/1_all_data_chunk.ipynb <==
# Split the large training data to several smaller csv file

來源： 玉山比賽 slack 群組官方問答區 channel 陳淳澐於 2021/11/23 提供 pandasDATAchunk.py 檔案修改而成

## import package

import pandas as pd 
import os
from tqdm import tqdm

## function defining

def reader_pandas_datacount(file, chunkSize):
    reader = pd.read_csv(file, chunksize=chunkSize)
    
    totalData=0
    for chunk in reader:
        chunkData_total = chunk.shape
        patitions = chunkData_total[0]
        
        with tqdm(range(patitions), 'Reading......') as t:
            for _ in t:
                try:
                    totalData = totalData + 1       
                    
                except StopIteration:
                        break
        print("\nchunk : " + str(totalData))
        print("data count : " + str(totalData*chunkSize))
                
    print("料估計完畢")
    print("總共 " + str(totalData*chunkSize) + " 筆資料")
    input("請按 Enter 繼續")
    return chunkSize, totalData

# Progressbar進度讀取條功能，此為個人修改後的版本 #

def Myreader_pandas(file, chunkSize, memoryflag=False):
    
    print("資料估計中......")
    chunkSize, patitions_total = reader_pandas_datacount(file,chunkSize)
    
    print("資料讀取中......")
    reader = pd.read_csv(file, iterator=True)
    chunks = []     
    
    j_No=0
    with tqdm(range(patitions_total), 'Reading......') as t:
        for _ in t:
            try:
                chunk = reader.get_chunk(chunkSize)
                #chunks.append(chunk)                
                #df_feature = pd.concat(chunks, ignore_index=True)
                #df_feature = pd.concat(chunk, ignore_index=True)
        
                if memoryflag == False:
                    # 建立存檔用料夾(名稱: 變數FolderName)
                    FolderName = "Tbrain_splitData"
                    savefileFolderPath = os.path.join("./",FolderName)
                
                    if not os.path.exists(savefileFolderPath):
                        os.mkdir(savefileFolderPath)
                    
                    # 存檔: .csv檔     
                    fileName = "tbrain_cc_SplitTraining_data_part" + str(j_No) + ".csv"            
                    savefilePath = os.path.join(savefileFolderPath,fileName)                
                
                    #df_feature.to_csv(savefilePath, mode='a')
                    chunk.to_csv(savefilePath, mode='a')
                j_No = j_No + 1        
            except StopIteration:
                break    
            
    return pd.concat(chunks, ignore_index=True)

## Main function

RUNfilePath = os.path.join("../../","train")
DatafileName = "train.csv"
openfilePath = os.path.join(RUNfilePath,DatafileName)

# 資料拆分流程 #
# 讀取資料集，很大所以用chunkSize加速讀取時間    
tmp = Myreader_pandas(openfilePath, 100000)
tmp.info()
##tmp = Myreader_pandas(openfilePath, 100000,True)
