==> src/nn/2_count_how_many.ipynb <==
# To count top 3 most category in last two month

## Import packages

import numpy as np
import pandas as pd 
import os

## Main function

if __name__ == '__main__':
    count = np.zeros(49);
    for i in range(20):
        filePath = "./Tbrain_splitData/tbrain_cc_SplitTraining_data_part" + str(310+i) + ".csv"
        print(filePath)
        df = pd.read_csv("./Tbrain_splitData/tbrain_cc_SplitTraining_data_part" + str(i) + ".csv")
        data = df.values.tolist() #整個轉為list
        for j in range(len(data)):
            if(data[j][3]!= 'other'):
                count[int(data[j][3])] = count[int(data[j][3])] + 1
    print(count)

category = [2, 6, 10, 12, 13, 15, 18, 19, 21, 22, 25, 26, 36, 37, 39, 48]

## Get the most three index category

sort = np.argsort(count).tolist()
for itr in sort:
    in_category = False
    for j in category:
        if (itr == j):
            in_category = True
    if(not in_category):
        sort.remove(itr)

test_data = np.zeros((500000,4))
for i in range(500000):
    test_data[i][0] = 10000000 + i
    test_data[i][1] = sort[-1]
    test_data[i][2] = sort[-2]
    test_data[i][3] = sort[-3]
print(test_data)

## Output file

most_df=pd.DataFrame(test_data, columns = ["chid", "top1", "top2", "top3"])
most_df_str = most_df.astype({'chid':'int32', 'top1':'int32', 'top2':'int32', 'top3':'int32'})
most_df_str.to_csv("./result/test_data.csv", index=False)
