==> src/nn/3_count_most_label_customer.ipynb <==
# Count the three most label for 500000 customers 

## Import packages

# For data preprocess
import numpy as np
import csv
import math
import pandas as pd

## Count the three most labels of every customer

data = np.zeros((500000,16))
shopping_type = [2,6,10,12,13,15,18,19,21,22,25,26,36,37,39,48]
for j in range(15):
    print('begin ' + str(j) + ' iterations')
    num = 314 + j
    data_csv = './Tbrain_splitData/tbrain_cc_SplitTraining_data_part' + str(num) +'.csv'
    df = pd.read_csv(data_csv)
    for i in range(100000):
        for k in range(16):
            if(df['shop_tag'][i] != 'other'):
                if(int(df['shop_tag'][i]) == shopping_type[k]):
                    data[df['chid'][i] - 10000000][k] = data[df['chid'][i] - 10000000][k] + df['txn_amt'][i]
    print('finish ' + str(j) + ' iterations')

## Get three most index of every customer

customer_id = np.zeros((500000,1))
for i in range(500000):
    customer_id[i] = 10000000 + i

most = np.zeros((500000,3))
for i in range(500000):
    for j in range(3):
        most[i][j] = -1
data_sort = np.argsort(data)    
for i in range(500000):
    for j in range(3):
        if(data[i][data_sort[i][-1*j-1]]!=0):
            most[i][j] = shopping_type[data_sort[i][-1*j-1]]

total = np.concatenate((customer_id, most), axis=1)

## Output file

most_df=pd.DataFrame(total, columns = ["chid", "top1", "top2", "top3"])
most_df_str = most_df.astype({'chid':'int32', 'top1':'int32', 'top2':'int32', 'top3':'int32'})
most_df_str.to_csv("./data/all_record.csv", index=False)
