==> src/nn/4_classification.ipynb <==
# Classify the customer to different files

## Import packages

# For data preprocess
import numpy as np
import csv
import pandas as pd

## Save file for the customer which has three category

df = pd.read_csv('./data/all_record.csv')
indexNames = df[ (df['top3'] == -1) ].index
df.drop(indexNames , inplace=True)
df.to_csv("./data/full_record.csv", index=False)

## Save file for the customer which has at least category but not three category

df = pd.read_csv('./data/all_record.csv')
indexNames = df[ (df['top3'] != -1) | (df['top1'] == -1) ].index
df.drop(indexNames , inplace=True)
df.to_csv("./data/not_full_record.csv", index=False)

## Save file for the customer which has no category

df = pd.read_csv('./data/all_record.csv')
indexNames = df[ (df['top1'] != -1)].index
df.drop(indexNames , inplace=True)
df.to_csv("./data/empty_record.csv", index=False)
