==> src/nn/5_reduce_training_data.ipynb <==
# Reduce the training data to 9 columns

## Import packages

# For data preprocess
import numpy as np
import csv
import math
import pandas as pd

## Main Function

data = np.zeros((500000,9))
for j in range(15):
    print('begin ' + str(j) + ' iterations')
    num = 314 + j
    data_csv = './Tbrain_splitData/tbrain_cc_SplitTraining_data_part' + str(num) +'.csv'
    df = pd.read_csv(data_csv)
    for i in range(100000):
        data[df['chid'][i] - 10000000][0] = df['masts'][i]
        data[df['chid'][i] - 10000000][1] = df['educd'][i]
        data[df['chid'][i] - 10000000][2] = df['trdtp'][i]
        data[df['chid'][i] - 10000000][3] = df['naty'][i]
        data[df['chid'][i] - 10000000][4] = df['poscd'][i]
        data[df['chid'][i] - 10000000][5] = df['cuorg'][i]
        data[df['chid'][i] - 10000000][6] = df['slam'][i]
        data[df['chid'][i] - 10000000][7] = df['gender_code'][i]
        data[df['chid'][i] - 10000000][8] = df['age'][i]
    print('finish ' + str(j) + ' iterations')

## Save file

customer_id = np.zeros((500000,1))
for i in range(500000):
    customer_id[i] = 10000000 + i
total = np.concatenate((customer_id, data), axis=1)


total_df=pd.DataFrame(total, columns = ["chid", "masts", "educd", "trdtp", "naty", "poscd", "cuorg", "slam", "gender_code","age"])


total_df=pd.DataFrame(total, columns = ["chid", "masts", "educd", "trdtp", "naty", "poscd", "cuorg", "slam", "gender_code","age"])
for i in range(len(total_df)):
    if(total_df["slam"][i] > 1000000):
        total_df["slam"][i] = 1000000
total_df=total_df.fillna(total_df.mean())
total_df.to_csv("./data/reduced_train.csv", index=False)
