==> src/nn/6_split_reduced_train.ipynb <==
# Split reduced training data to two partially part

## Import package

# For data preprocess
import numpy as np
import csv
import pandas as pd

## Import csv file

df = pd.read_csv('./data/reduced_train.csv')
df_full = pd.read_csv('./data/full_record.csv')
not_df_full = pd.read_csv('./data/not_full_record.csv')

## Get chid & whether it is full

id_full = df_full['chid'].tolist()
id_not_full = not_df_full['chid'].tolist()

full_train_df = df[df.chid.isin(id_full)]
not_full_train_df = df[df.chid.isin(id_not_full)]

## Output file

full_train_df.to_csv("./data/full_reduced_train.csv", index=False)
not_full_train_df.to_csv("./data/not_full_reduced_train.csv", index=False)
