==> src/nn/8_output_result.ipynb <==
# Output result

## Import packages

# For data preprocess
import numpy as np
import csv
import math
import pandas as pd

## Combine different csv file to generate total csv file 

data = np.zeros((500000,3))
df = pd.read_csv('./data/empty_record.csv')
for i in range(len(df)):
    data[int(df['chid'][i])-10000000][0] = 37
    data[int(df['chid'][i])-10000000][1] = 15
    data[int(df['chid'][i])-10000000][2] = 2
df = pd.read_csv('./data/full_record.csv')
for i in range(len(df)):
    data[int(df['chid'][i])-10000000][0] = df['top1'][i]
    data[int(df['chid'][i])-10000000][1] = df['top2'][i]
    data[int(df['chid'][i])-10000000][2] = df['top3'][i]
df = pd.read_csv('./data/predict_not_full.csv')
for i in range(len(df)):
    data[int(df['chid'][i])-10000000][0] = df['top1'][i]
    data[int(df['chid'][i])-10000000][1] = df['top2'][i]
    data[int(df['chid'][i])-10000000][2] = df['top3'][i]

## Output file

with open('./result/result.csv', 'w', newline='') as csvfile:
    writer = csv.writer(csvfile)
    writer.writerow(['chid', 'top1','top2','top3'])
    for i in range(len(data)):
        writer.writerow([str(i + 10000000), str(int(data[i][0])),str(int(data[i][1])),str(int(data[i][2]))])
