==> src/nn/9_check_result.ipynb <==
# Check the format of predicted result

import os
import csv
import numpy as np
import pandas as pd

## Read Data

df = pd.read_csv("./result/result.csv")

train_chid = df['chid'];
train_top1 = df['top1'];
train_top2 = df['top2'];
train_top3 = df['top3'];

## To verify that top 3 label are different

passSignal = True
for i in range (500000):
    if(train_top1[i] == train_top2[i]):
        passSignal = False;
        print("something wrong! ", i, "th data")
        print(train_chid[i])
    if(train_top1[i] == train_top3[i]):
        passSignal = False;
        print("something wrong! ", i, "th data")
        print(train_chid[i])
    if(train_top2[i] == train_top3[i]):
        passSignal = False;
        print("something wrong! ", i, "th data")
        print(train_chid[i])
if(passSignal):
    print("Congratulation! Top 3 label are totally different")

## To verify that every label is in the predicted label list

passSignal = True
label = [2,6,10,12,13,15,18,19,21,22,25,26,36,37,39,48]
for i in range (500000):
    if(not i%100000):
        print("begin ",i," turns")
    sameSignal = False;
    for j in range(16):
        if(train_top1[i] == label[j]):
            sameSignal = True;
    if(not sameSignal):
        print("something wrong! ", i, "th data")
        passSignal = False;
    
    sameSignal = False;
    for j in range(16):
        if(train_top2[i] == label[j]):
            sameSignal = True;
    if(not sameSignal):
        print("something wrong! ", i, "th data")
        passSignal = False;
    
    sameSignal = False;
    for j in range(16):
        if(train_top3[i] == label[j]):
            sameSignal = True;
    if(not sameSignal):
        print("something wrong! ", i, "th data")
        passSignal = False;
if(passSignal):
    print("Congratulation! Every label is in the predicted label list.")

## To check the type of data
### chid: int64
### top1: int64
### top2: int64
### top3: int64
### dtype: object

print(df.dtypes)
